[BUGFIX] Resolves #11082: Removed the Tx_ExtbaseKickstarter_Utility_Compatability class
This class made a check based on versions which did seem to cause problems
in the current situation. Since the real difference between those versions
are the existence of a class or the availability of depency injection
I rewrote the constructors which makes the current trunk compatible
with TYPO3 4.4.5 release and current trunk of TYPO3.