[TASK] Added support for harvesting template variables, ExposedTemplateView
This change adds a fallback in case there is no root variablecontainer value 
for the variable you are harvesting - in which case the ExposedTemplateView now 
looks for a template variable of the same name and returns it if it exists, else NULL 
as before this change.

Needless to say, root variablecontainer variables still have priority over 
templatecontainer variables.